pyproject.toml
[project]
name = "python-repo-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.8",
    "mlflow>=3.8.1",
    "numpy>=2.4.1",
    "omegaconf>=2.3.0",
    "pandas>=2.3.3",
    "rich>=14.3.1",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.pytest.ini_options]
pythonpath = [
    "src"
]